feat: esm.js solve
